break apart validator helpers
