+++ src/shared/api/OneSignalApiBase.ts
@@ -1,6 +1,7 @@
 import { getOneSignalApiUrl } from '../environment/detect';
 import { AppIDMissingError, RetryLimitError } from '../errors/common';
-import { delay, isValidUuid } from '../helpers/general';
+import { delay } from '../helpers/general';
+import { isValidUuid } from '../helpers/validators';
 import Log from '../libraries/Log';
 import type { APIHeaders } from '../models/APIHeaders';
 import { IS_SERVICE_WORKER, VERSION } from '../utils/EnvVariables';
