+++ src/shared/helpers/MainHelper.ts
@@ -1,6 +1,8 @@
 import type { NotificationIcons } from 'src/shared/notifications/types';
-import { ValidatorUtils } from '../../page/utils/ValidatorUtils';
-import { getOneSignalApiUrl, useSafariLegacyPush } from '../environment/detect';
+import {
+  getOneSignalApiUrl,
+  useSafariLegacyPush,
+} from '../environment/detect';
 import { AppIDMissingError, MalformedArgumentError } from '../errors/common';
 import Log from '../libraries/Log';
 import type {
@@ -11,6 +13,7 @@ import Database from '../services/Database';
 import { PermissionUtils } from '../utils/PermissionUtils';
 import { getPlatformNotificationIcon, logMethodCall } from '../utils/utils';
 import { getValueOrDefault } from './general';
+import { isValidUrl } from './validators';
 
 export default class MainHelper {
   static async showLocalNotification(
@@ -36,10 +39,8 @@ export default class MainHelper {
     if (!appConfig.appId) throw AppIDMissingError;
     if (!OneSignal.Notifications.permission)
       throw new Error('User is not subscribed');
-    if (!ValidatorUtils.isValidUrl(url)) throw MalformedArgumentError('url');
-    if (
-      !ValidatorUtils.isValidUrl(icon, { allowEmpty: true, requireHttps: true })
-    )
+    if (!isValidUrl(url)) throw MalformedArgumentError('url');
+    if (!isValidUrl(icon, { allowEmpty: true, requireHttps: true }))
       throw MalformedArgumentError('icon');
     if (!icon) {
       // get default icon
